Product/requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
appdirs==1.4.4
asgiref==3.11.0
astunparse==1.6.3
attrs==25.3.0
audioread==3.0.1
av==16.0.1
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.3.0
cachetools==6.2.2
certifi==2025.8.3
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.3
chroma-hnswlib==0.7.3
chromadb==1.1.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
coqpit-config==0.2.1
coqui-tts==0.27.1
coqui-tts-trainer==0.3.1
crewai==1.5.0
cryptography==46.0.3
ctranslate2==4.6.1
cycler==0.12.1
Cython==3.1.4
dataclasses-json==0.6.7
dateparser==1.1.8
decorator==5.2.1
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docopt==0.6.2
docstring_parser==0.17.0
dotenv==0.9.9
durationpy==0.10
einops==0.8.1
email-validator==2.3.0
encodec==0.1.1
et_xmlfile==2.0.0
fastapi==0.117.1
faster-whisper==1.2.1
fastuuid==0.14.0
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
gast==0.6.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.52.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
groq==0.36.0
grpcio==1.75.0
grpcio-status==1.71.2
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.0
humanfriendly==10.0
identify==2.6.15
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
instructor==1.13.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
json5==0.12.1
json_repair==0.25.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.0.8
langchain-chroma==0.1.4
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-google-genai==3.1.0
langchain-groq==1.0.1
langchain-huggingface==0.0.1
langchain-ollama==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.9
langsmith==0.4.46
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
litellm==1.80.5
llvmlite==0.45.0
loguru==0.7.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mcp==1.22.0
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
monotonic-alignment-search==0.2.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.5
nodeenv==1.9.1
nomic==3.9.0
num2words==0.5.14
numba==0.62.0
numpy==1.26.4
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai==2.8.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==23.2
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.5.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.4
PyMuPDF==1.26.6
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==2.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
SpeechRecognition==3.14.4
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.48.0
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
text-generation==0.7.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
transformers==4.41.2
twilio==9.8.1
ty==0.0.1a27
typeguard==4.4.4
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
uv==0.9.11
uvicorn==0.37.0
virtualenv==20.35.4
watchfiles==1.1.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
